+++ persistence/PostgresPersistence.go
@@ -20,7 +20,15 @@ import (
 	cref "github.com/pip-services3-go/pip-services3-commons-go/refer"
 	clog "github.com/pip-services3-go/pip-services3-components-go/log"
 	cmpersist "github.com/pip-services3-go/pip-services3-data-go/persistence"
+	conn "github.com/pip-services3-go/pip-services3-postgres-go/connect"
 )
+
+type IPostgresPersistenceOverrides interface {
+	DefineSchema()
+	ConvertFromPublic(item interface{}) interface{}
+	ConvertToPublic(item pgx.Rows) interface{}
+	ConvertFromPublicPartial(item interface{}) interface{}
+}
 
 /*
 Abstract persistence component that stores data in PostgreSQL using plain driver.
@@ -92,6 +100,9 @@ accessing c._db or c._collection properties.
 */
 
 type PostgresPersistence struct {
+	Overrides IPostgresPersistenceOverrides
+	Prototype reflect.Type
+
 	defaultConfig *cconf.ConfigParams
 
 	config           *cconf.ConfigParams
@@ -100,17 +111,12 @@ type PostgresPersistence struct {
 	localConnection  bool
 	schemaStatements []string
 
-	ConvertFromPublic        func(interface{}) interface{}
-	ConvertToPublic          func(pgx.Rows) interface{}
-	ConvertFromPublicPartial func(interface{}) interface{}
-	DefineSchema             func()
-
 	//The dependency resolver.
 	DependencyResolver *cref.DependencyResolver
 	//The logger.
 	Logger *clog.CompositeLogger
 	//The PostgreSQL connection component.
-	Connection *PostgresConnection
+	Connection *conn.PostgresConnection
 	//The PostgreSQL connection pool object.
 	Client *pgxpool.Pool
 	//The PostgreSQL database name.
@@ -118,13 +124,15 @@ type PostgresPersistence struct {
 	//The PostgreSQL table object.
 	TableName   string
 	MaxPageSize int
-	Prototype   reflect.Type
 }
 
 // Creates a new instance of the persistence component.
+//   - overrides References to override virtual methods
 //   - tableName    (optional) a table name.
-func NewPostgresPersistence(proto reflect.Type, tableName string) *PostgresPersistence {
+func InheritPostgresPersistence(overrides IPostgresPersistenceOverrides, proto reflect.Type, tableName string) *PostgresPersistence {
 	c := &PostgresPersistence{
+		Overrides: overrides,
+		Prototype: proto,
 		defaultConfig: cconf.NewConfigParamsFromTuples(
 			"collection", nil,
 			"dependencies.connection", "*:connection:postgres:*:1.0",
@@ -138,16 +146,12 @@ func NewPostgresPersistence(proto reflect.Type, tableName string) *PostgresPersi
 		schemaStatements: make([]string, 0),
 		Logger:           clog.NewCompositeLogger(),
 		MaxPageSize:      100,
-		Prototype:        proto,
+		TableName:        tableName,
 	}
-
-	c.ConvertFromPublic = c.PerformConvertFromPublic
-	c.ConvertToPublic = c.PerformConvertToPublic
-	c.ConvertFromPublicPartial = c.PerformConvertFromPublic
 
 	c.DependencyResolver = cref.NewDependencyResolver()
 	c.DependencyResolver.Configure(c.defaultConfig)
-	c.TableName = tableName
+
 	return c
 }
 
@@ -173,7 +177,7 @@ func (c *PostgresPersistence) SetReferences(references cref.IReferences) {
 	// Get connection
 	c.DependencyResolver.SetReferences(references)
 	result := c.DependencyResolver.GetOneOptional("connection")
-	if dep, ok := result.(*PostgresConnection); ok {
+	if dep, ok := result.(*conn.PostgresConnection); ok {
 		c.Connection = dep
 	}
 	// Or create a local one
@@ -190,8 +194,8 @@ func (c *PostgresPersistence) UnsetReferences() {
 	c.Connection = nil
 }
 
-func (c *PostgresPersistence) createConnection() *PostgresConnection {
-	connection := NewPostgresConnection()
+func (c *PostgresPersistence) createConnection() *conn.PostgresConnection {
+	connection := conn.NewPostgresConnection()
 	if c.config != nil {
 		connection.Configure(c.config)
 	}
@@ -238,6 +242,11 @@ func (c *PostgresPersistence) EnsureIndex(name string, keys map[string]string, o
 	c.EnsureSchema(builder)
 }
 
+// Defines a database schema for this persistence
+func (c *PostgresPersistence) DefineSchema() {
+	// Override in child classes
+}
+
 // // Adds a statement to schema definition.
 // // This is a deprecated method. Use EnsureSchema instead.
 // //   - schemaStatement a statement to be added to the schema
@@ -259,8 +268,7 @@ func (c *PostgresPersistence) ClearSchema() {
 // Converts object value from internal to func (c * PostgresPersistence) format.
 //   - value     an object in internal format to convert.
 // Returns converted object in func (c * PostgresPersistence) format.
-func (c *PostgresPersistence) PerformConvertToPublic(rows pgx.Rows) interface{} {
-
+func (c *PostgresPersistence) ConvertToPublic(rows pgx.Rows) interface{} {
 	values, valErr := rows.Values()
 	if valErr != nil || values == nil {
 		return nil
@@ -282,14 +290,14 @@ func (c *PostgresPersistence) PerformConvertToPublic(rows pgx.Rows) interface{} 
 // Convert object value from func (c * PostgresPersistence) to internal format.
 //   - value     an object in func (c * PostgresPersistence) format to convert.
 // Returns converted object in internal format.
-func (c *PostgresPersistence) PerformConvertFromPublic(value interface{}) interface{} {
+func (c *PostgresPersistence) ConvertFromPublic(value interface{}) interface{} {
 	return value
 }
 
 // Converts the given object from the public partial format.
 //   - value     the object to convert from the public partial format.
 // Returns the initial object.
-func (c *PostgresPersistence) PerformConvertFromPublicPartial(value interface{}) interface{} {
+func (c *PostgresPersistence) ConvertFromPublicPartial(value interface{}) interface{} {
 	return c.ConvertFromPublic(value)
 }
 
@@ -343,9 +351,7 @@ func (c *PostgresPersistence) Open(correlationId string) (err error) {
 	c.DatabaseName = c.Connection.GetDatabaseName()
 
 	// Define database schema
-	if c.DefineSchema != nil {
-		c.DefineSchema()
-	}
+	c.Overrides.DefineSchema()
 
 	// Recreate objects
 	err = c.CreateSchema(correlationId)
@@ -616,7 +622,7 @@ func (c *PostgresPersistence) GetPageByFilter(correlationId string, filter inter
 
 	items := make([]interface{}, 0, 0)
 	for qResult.Next() {
-		item := c.ConvertToPublic(qResult)
+		item := c.Overrides.ConvertToPublic(qResult)
 		items = append(items, item)
 	}
 
@@ -725,7 +731,7 @@ func (c *PostgresPersistence) GetListByFilter(correlationId string, filter inter
 	defer qResult.Close()
 	items = make([]interface{}, 0, 1)
 	for qResult.Next() {
-		item := c.ConvertToPublic(qResult)
+		item := c.Overrides.ConvertToPublic(qResult)
 		items = append(items, item)
 	}
 
@@ -787,7 +793,7 @@ func (c *PostgresPersistence) GetOneRandom(correlationId string, filter interfac
 		c.Logger.Trace(correlationId, "Random item wasn't found from %s", c.TableName)
 		return nil, qResult2.Err()
 	}
-	item = c.ConvertToPublic(qResult2)
+	item = c.Overrides.ConvertToPublic(qResult2)
 	c.Logger.Trace(correlationId, "Retrieved random item from %s", c.TableName)
 	return item, nil
 
@@ -803,7 +809,7 @@ func (c *PostgresPersistence) Create(correlationId string, item interface{}) (re
 		return nil, nil
 	}
 
-	row := c.ConvertFromPublic(item)
+	row := c.Overrides.ConvertFromPublic(item)
 	columns := c.GenerateColumns(row)
 	params := c.GenerateParameters(row)
 	values := c.GenerateValues(columns, row)
@@ -816,7 +822,7 @@ func (c *PostgresPersistence) Create(correlationId string, item interface{}) (re
 	if !qResult.Next() {
 		return nil, qResult.Err()
 	}
-	item = c.ConvertToPublic(qResult)
+	item = c.Overrides.ConvertToPublic(qResult)
 	id := cmpersist.GetObjectId(item)
 	c.Logger.Trace(correlationId, "Created in %s with id = %s", c.TableName, id)
 	return item, nil